// FrontEnd/src/Admin/models/ProductModels.ts
export interface ProductModel {
    id: string;
    name: string;
    description: string;
    weight: string;
    size: string;
    unit: string;
    quantity: string;
    price: string;
    price_sale: string;
    category_id: string;
    category_name: string;
    thumbnail: string;
    slug: string;
    created_at: string;
    updated_at: string;
    modified_by: string;
    deleted_at?: any;
}